Ignore doc test
In one of the documentation examples I added in a previous commit, the
code refers to a local file `somebinary.exe`, which does not exist in
the repository. Therefore, as Rust runs documentation examples as tests,
this causes the test suite to fail when run. This commit ignores this
example, so it will not cause the test suite to fail.