Add `scan_first_match` function.
This function is used to find the index of the first location in a byte
string which matches a pattern. This should be more efficient if you
just want to test if a byte string matches a signature, or if you just
want the first index of a match, as it can take advantage of an early
return as soon as a match is found, rather than continuing to find
matches throughout the byte string.